Changed int types to u8

+++ bitboard.cpp
@@ -22,14 +22,14 @@ bitboard::~bitboard()
 u64 bitboard::getBoard()
 {
 	u64 bboard = 0;
-	for (int i = 0; i < 12; i++)
+	for (u8 i = 0; i < 12; i++)
 		bboard|=board[i];
 	return bboard;
 }
 
 void bitboard::setBoard(u64 bboard[])
 {
-	for (int i = 0; i < 12; i++)
+	for (u8 i = 0; i < 12; i++)
 		board[i] = bboard[i];
 	return;
 }
